alexnet_API/requirements.txt
absl-py==0.7.1
astor==0.7.1
certifi==2019.3.9
Django==2.2.1
django-cors-headers==2.5.3
djangorestframework==3.9.3
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
mock==3.0.4
numpy==1.16.3
Pillow==6.0.0
protobuf==3.7.1
pytz==2019.1
PyYAML==5.1
scipy==1.2.1
six==1.12.0
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.2
